Add a tester method to get the current peer name.
The current peer name is the IP of the browser currently connected
to the web UI. This is no constant value, as a user can disconnect
and allow another user with a different IP to connect instead.
It is implemented as a request from the test to the web server.