Read system parameter from the tests.xml .
System parameter are defined like test parameter, just as children
of "System":

  <?xml version="1.0" encoding="UTF-8"?>
  <MuhkuhTest title="netRAPID R4" subtitle="7690.103 R4">
          <System>
                  <Parameter name="manufacturer">1</Parameter>
                  <Parameter name="devicenr">7690103</Parameter>
                  <Parameter name="hwrev">4</Parameter>
          </System>
  ...

Please note that system parameter can only be defined with the "Parameter"
element. It is not possible to use "Connection" as a child of "System".

Please note that the system parameter are automatically extendet by the
tester system with the current serial under test as "serial".

In a test case it is possible to make connections to system parameters:

  ...
          <Testcase file="../../local/test_patch_fdl.lua" name="Write_FDL">
                <Parameter name="group">HilscherDefault</Parameter>
                <Connection name="manufacturer">system:manufacturer</Connection>
                <Connection name="devicenr">system:devicenr</Connection>
                <Connection name="serial">system:serial</Connection>
                <Connection name="hwrev">system:hwrev</Connection>
  ...